rename timestamp components for clarity
